+++ src/AlarmClock/DaysOfWeek.h
@@ -22,6 +22,14 @@ protected:
 
 
 public:
+    /*!
+        @brief  A one byte long representation of the object.
+        The value for each day of the week is represented as a single bit in
+        this number.
+        Bit 0 (LSB) means nothing,
+        bit 1 is Monday,
+        bit 7 (MSB) is Sunday.
+    */
     byte DaysOfWeek;
     bool getDayOfWeek(byte num) const;
     bool getDayOfWeek_Adafruit(byte num) const;