+++ src/AlarmClock/Settings.h
@@ -159,8 +159,13 @@ I2C addresses
 /*
 Serial CLI
 */
-// Width of indentation (e.g. in help)
+// Width of indentation (e.g. in help and YAML)
 #define Serial_indentation_width 2  // characters
+
+// YAML_begin and YAML_end need to be put on separate lines with no other
+// characters (except for '\r' and '\n')
+#define YAML_begin F("---")
+#define YAML_end F("...")
 
 // Time elapsed since last command after which the CLI writes the changed data
 // to the EEPROM