+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Incompatibility with new RTClib
 
+### Changed
+- CLI rework. It should be much easier to parse. Commands now output YAML.  
+  **WARNING**: breaking CLI changes
+
 
 ## [0.4.1] - 2021-03-11
 ### Fixed